+++ changelog.md
@@ -15,7 +15,7 @@
 	- All Bullet Kin are despawned at the end of the pizza minigame, rather than just the ones without pizza
 	- Adjusted Bullet Kin placement to better avoid being hidden by walls
 	- Chest rewards from the pizza minigame now spawn unlocked
-- Added Nightlighter
+- Added Nightlighter and Gadulo
 - Added mastery for Nightlighter
 - Added 2 new synergies